split up ui.py and created ui_elements.py
